social_app/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-filter==21.1
djangorestframework==3.12.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
Markdown==3.3.6
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
